+++ WebApp/Controllers/SubscriptionPlansController.cs
@@ -2,33 +2,69 @@
 using LightMethods.Survey.Models.DTO;
 using LightMethods.Survey.Models.Entities;
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Web.Http;
+using WebApi.Results;
+using WebApi.Services;
 
 namespace WebApi.Controllers
 {
     [RoutePrefix("api/subscriptionplans")]
+    [Authorize(Roles = "System administrator,Platform administrator")]
     public class SubscriptionPlansController : BaseApiController
     {
+
+        private const string CACHE_KEY = "SUBSCRIPTION_PLANS";
 
         // GET api/subscriptionPlans
         public IHttpActionResult Get()
         {
-            var plans = UnitOfWork.SubscriptionPlansRepository.AllAsNoTracking;
-            var result = plans.ToList().Select(sp => Mapper.Map<SubscriptionPlanDTO>(sp));
+            var cacheEntry = MemoryCacher.GetValue(CACHE_KEY);
+            if (cacheEntry == null)
+            {
+                var plans = UnitOfWork.SubscriptionPlansRepository
+                    .AllAsNoTracking
+                    .ToList();
 
-            return Ok(result);
+                var result = plans.Select(sp => Mapper.Map<SubscriptionPlanDTO>(sp)).ToList();
+                MemoryCacher.Add(CACHE_KEY, result, DateTimeOffset.UtcNow.AddMinutes(1));
+
+                return Ok(result);
+            }
+            else
+            {
+                var result = (List<SubscriptionPlanDTO>)cacheEntry;
+                return new CachedResult<List<SubscriptionPlanDTO>>(result, TimeSpan.FromMinutes(1), this);
+            }
         }
 
         // GET api/subscriptionPlans/{id}
         [Route("{id:guid}")]
         public IHttpActionResult Get(Guid id)
         {
-            var plan = UnitOfWork.SubscriptionPlansRepository.Find(id);
-            if (plan == null)
-                return NotFound();
+            if (id == Guid.Empty)
+                return BadRequest("id is empty");
 
-            return Ok(Mapper.Map<SubscriptionPlanDTO>(plan));
+            var cacheKey = $"{CACHE_KEY}_{id}";
+            var cacheEntry = MemoryCacher.GetValue(cacheKey);
+
+            if (cacheEntry == null)
+            {
+                var plan = UnitOfWork.SubscriptionPlansRepository.Find(id);
+                if (plan == null)
+                    return NotFound();
+
+                var result = Mapper.Map<SubscriptionPlanDTO>(plan);
+                MemoryCacher.Add(cacheKey, result, DateTimeOffset.UtcNow.AddMinutes(1));
+
+                return Ok(result);
+            }
+            else
+            {
+                var result = (SubscriptionPlanDTO)cacheEntry;
+                return new CachedResult<SubscriptionPlanDTO>(result, TimeSpan.FromMinutes(1), this);
+            }
         }
 
         // POST api/subscriptionPlans
@@ -36,16 +72,19 @@ namespace WebApi.Controllers
         public IHttpActionResult Post([FromBody]SubscriptionPlanDTO value)
         {
             var plan = Mapper.Map<SubscriptionPlan>(value);
+
             try
             {
                 UnitOfWork.SubscriptionPlansRepository.InsertOrUpdate(plan);
                 UnitOfWork.Save();
 
+                MemoryCacher.Delete(CACHE_KEY);
+
                 return Ok();
             }
             catch (Exception ex)
             {
-                return BadRequest(ex.Message);
+                return InternalServerError(ex);
             }
         }
 
@@ -55,7 +94,7 @@ namespace WebApi.Controllers
         public IHttpActionResult Put(Guid id, [FromBody]SubscriptionPlanDTO value)
         {
             if (id == Guid.Empty)
-                return BadRequest();
+                return BadRequest("id is empty");
 
             var plan = UnitOfWork.SubscriptionPlansRepository.Find(id);
             if (plan == null)
@@ -75,11 +114,13 @@ namespace WebApi.Controllers
                 UnitOfWork.SubscriptionPlansRepository.InsertOrUpdate(plan);
                 UnitOfWork.Save();
 
+                MemoryCacher.DeleteListAndItem(CACHE_KEY, id);
+
                 return Ok();
             }
             catch (Exception ex)
             {
-                return BadRequest(ex.Message);
+                return InternalServerError(ex);
             }
         }
 
@@ -88,17 +129,23 @@ namespace WebApi.Controllers
         [Route("{id:guid}")]
         public IHttpActionResult Delete(Guid id)
         {
+            if (id == Guid.Empty)
+                return BadRequest("id is empty");
+
             try
             {
                 UnitOfWork.SubscriptionPlansRepository.Delete(id);
                 UnitOfWork.Save();
 
+                MemoryCacher.DeleteListAndItem(CACHE_KEY, id);
+
                 return Ok();
             }
             catch (Exception ex)
             {
-                return BadRequest(ex.Message);
+                return InternalServerError(ex);
             }
         }
+
     }
 }
